+++ src/util/schemas.js
@@ -9,8 +9,7 @@ module.exports = {
   import: Joi.object({
     accessKey: Joi.any().valid(`Bearer ${config.accessKey}`).required(),
     cookies: Joi.string().required(),
-    url:
-      Joi.string().pattern(new RegExp('^https://exhentai.org.+')).required(),
+    url: Joi.string().pattern(/^https:\/\/exhentai.org.+/).required(),
     skipImport: Joi.boolean(),
     skipKnownFiles: Joi.boolean(),
     deleteArchivesAfterImport: Joi.boolean(),
@@ -18,6 +17,7 @@ module.exports = {
     blacklistedNamespaces: Joi.string().allow(''),
     namespaceReplacements: Joi.string().allow(''),
     additionalTags: Joi.string().allow(''),
-    addUniqueIdentifierTag: Joi.boolean()
+    addUniqueIdentifierTag: Joi.boolean(),
+    uniqueIdentifierNamespace: Joi.string().allow('')
   })
 }